Added rotated text and improved clear_screen()
In screen_lib.py replaced the Screen.text() function with a new
version that rotates text by 90deg by default. This gives text that
is the right way up in the long direction of the screen.

clear_screen() improved by resetting the image to blank white and
performing 2 updates to populate both screen memories.

 On branch master
 Changes to be committed:
	modified:   screen_lib.py
	modified:   waveshare_epd_lib.py